오류 수정, 백엔드 안정화

+++ backend/src/main/java/com/team/shopping/Services/MultiService.java
@@ -1,13 +1,6 @@
 package com.team.shopping.Services;
 
 
-import com.team.shopping.DTOs.AuthRequestDTO;
-import com.team.shopping.DTOs.AuthResponseDTO;
-import com.team.shopping.DTOs.CategoryRequestDTO;
-import com.team.shopping.DTOs.SignupRequestDTO;
-import com.team.shopping.Domains.Auth;
-import com.team.shopping.Domains.Category;
-import com.team.shopping.Domains.SiteUser;
 import com.team.shopping.DTOs.*;
 import com.team.shopping.Domains.*;
 import com.team.shopping.Enums.UserRole;
@@ -15,18 +8,16 @@ import com.team.shopping.Exceptions.DataDuplicateException;
 import com.team.shopping.Records.TokenRecord;
 import com.team.shopping.Securities.CustomUserDetails;
 import com.team.shopping.Securities.JWT.JwtTokenProvider;
-import com.team.shopping.Services.Module.AuthService;
-import com.team.shopping.Services.Module.CategoryService;
-import com.team.shopping.Services.Module.ProductService;
-import com.team.shopping.Services.Module.UserService;
-import com.team.shopping.Services.Module.WishListService;
+import com.team.shopping.Services.Module.*;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -97,9 +88,7 @@ public class MultiService {
     public SiteUser signup(SignupRequestDTO signupRequestDTO) throws DataDuplicateException {
         userService.check(signupRequestDTO);
         SiteUser user = userService.save(signupRequestDTO);
-        Wish wishList = Wish.builder()
-                .user(user)
-                .build();
+        Wish wishList = Wish.builder().user(user).build();
         this.wishListService.save(wishList);
         return user;
     }
@@ -107,18 +96,7 @@ public class MultiService {
     @Transactional
     public UserResponseDTO getProfile(String username) {
         SiteUser siteUser = this.userService.get(username);
-        return UserResponseDTO.builder()
-                .username(siteUser.getUsername())
-                .gender(siteUser.getGender().toString())
-                .email(siteUser.getEmail())
-                .point(siteUser.getPoint())
-                .phoneNumber(siteUser.getPhoneNumber())
-                .nickname(siteUser.getNickname())
-                .birthday(siteUser.getBirthday())
-                .createDate(siteUser.getCreateDate())
-                .modifyDate(siteUser.getModifyDate())
-                .name(siteUser.getName())
-                .build();
+        return UserResponseDTO.builder().username(siteUser.getUsername()).gender(siteUser.getGender().toString()).email(siteUser.getEmail()).point(siteUser.getPoint()).phoneNumber(siteUser.getPhoneNumber()).nickname(siteUser.getNickname()).birthday(siteUser.getBirthday()).createDate(siteUser.getCreateDate()).modifyDate(siteUser.getModifyDate()).name(siteUser.getName()).build();
     }
 
     /**
@@ -164,6 +142,7 @@ public class MultiService {
         }
         this.productService.save(requestDTO, user);
     }
+
     @Transactional
     public void deleteUser(String username, String name) {
         SiteUser targetUser = userService.get(username);
@@ -179,12 +158,12 @@ public class MultiService {
      * Category
      */
     @Transactional
-    public void saveCategory(String username, CategoryRequestDTO requestDto) throws DataDuplicateException{
+    public void saveCategory(String username, CategoryRequestDTO requestDto) throws DataDuplicateException {
         SiteUser siteUser = userService.get(username);
-        if (siteUser.getRole().equals(UserRole.ADMIN)){
+        if (siteUser.getRole().equals(UserRole.ADMIN)) {
             this.categoryService.check(requestDto);
             this.categoryService.save(requestDto);
-        }else {
+        } else {
             throw new IllegalArgumentException("ADMIN 권한이 아닙니다.");
         }
 
@@ -193,35 +172,31 @@ public class MultiService {
     /**
      * 테스트용
      */
-
     @Transactional
-    public void updateCategory(CategoryRequestDTO requestDto) {
-        Optional<Category> _category = categoryService.get(requestDto.getId());
-        if (_category.isPresent()) {
-            categoryService.updateCheck(requestDto);
-            categoryService.update(_category.get(), requestDto.getNewName());
-        } else throw new IllegalArgumentException("해당 ID를 가진 카테고리가 존재하지 않습니다.");
-    }
+    public void deleteCategory(String username, Long id) {
+        SiteUser siteUser = userService.get(username);
+        if (siteUser.getRole().equals(UserRole.ADMIN)) {
+            Optional<Category> _category = categoryService.get(id);
+            if (_category.isPresent()) categoryService.deleteCategory(_category.get());
+            else throw new IllegalArgumentException("해당 ID를 가진 카테고리가 존재하지 않습니다.");
+        } else {
+            throw new IllegalArgumentException("ADMIN 권한이 아닙니다.");
+        }
 
-    @Transactional
-    public void deleteCategory(Long id) {
-        Optional<Category> _category = categoryService.get(id);
-        if (_category.isPresent())
-            categoryService.deleteCategory(_category.get());
-        else throw new IllegalArgumentException("해당 ID를 가진 카테고리가 존재하지 않습니다.");
     }
 
 
-//    @Transactional
-//    public void updateCategory(String username, CategoryRequestDTO requestDto) throws DataDuplicateException{
-//        SiteUser siteUser = userService.get(username);
-//        if (siteUser.getRole().equals(UserRole.ADMIN)){
-//            categoryService.updateCheck(requestDto);
-//            categoryService.update(requestDto);
-//        }else {
-//            throw new IllegalArgumentException("ADMIN 권한이 아닙니다.");
-//        }
-//    }
-
-
+    @Transactional
+    public void updateCategory(String username, CategoryRequestDTO requestDto) throws DataDuplicateException {
+        SiteUser siteUser = userService.get(username);
+        if (siteUser.getRole().equals(UserRole.ADMIN)) {
+            Optional<Category> _category = categoryService.get(requestDto.getId());
+            if (_category.isPresent()) {
+                categoryService.updateCheck(requestDto);
+                categoryService.update(_category.get(), requestDto.getNewName());
+            } else {
+                throw new IllegalArgumentException("ADMIN 권한이 아닙니다.");
+            }
+        }
+    }
 }
